+++ ufo/module/round.py
@@ -10,11 +10,13 @@ from ..automator.ui_control.screenshot import PhotographerFacade
 from ..config.config import Config
 from . import processor
 
+from .basic import BaseRound
+
 configs = Config.get_instance().config_data
 
 
 
-class Round(object):
+class Round(BaseRound):
     """
     A round of a session in UFO.
     """
@@ -71,6 +73,7 @@ class Round(object):
         self.AppAgent = self.HostAgent.get_active_appagent()
         self.app_window = host_agent_processor.get_active_window()
         self.application = host_agent_processor.get_active_control_text()
+        
 
 
     def process_action_selection(self) -> None:
@@ -89,83 +92,3 @@ class Round(object):
         self.update_cost(app_agent_processor.get_process_cost())
 
         self.control_reannotate = app_agent_processor.get_control_reannotate()
-
-
-    def get_status(self) -> str:
-        """
-        Get the status of the session.
-        return: The status of the session.
-        """
-        return self._status
-
-
-
-    def get_step(self) -> int:
-        """
-        Get the step of the session.
-        return: The step of the session.
-        """
-        return self._step
-
-
-    def get_cost(self) -> float:
-        """
-        Get the cost of the session.
-        return: The cost of the session.
-        """
-        return self._cost
-
-
-    def print_cost(self) -> None:
-        # Print the total cost 
-
-        total_cost = self.get_cost()  
-        if isinstance(total_cost, float):  
-            formatted_cost = '${:.2f}'.format(total_cost)  
-            utils.print_with_color(f"Request total cost for current round is {formatted_cost}", "yellow")
-
-
-    def get_results(self) -> str:
-        """
-        Get the results of the session.
-        return: The results of the session.
-        """
-
-        action_history = self.HostAgent.get_global_action_memory().content
-
-        if len(action_history) > 0:
-            result = action_history[-1].to_dict().get("Results")
-        else:
-            result = ""
-        return result
-
-
-    def set_index(self, index: int) -> None:
-        """
-        Set the index of the session.
-        """
-        self.index = index
-
-    def set_global_step(self, global_step: int) -> None:
-        """
-        Set the global step of the session.
-        """
-        self.global_step = global_step
-
-
-    def get_application_window(self) -> object:
-        """
-        Get the application of the session.
-        return: The application of the session.
-        """
-        return self.app_window
-
-
-    def update_cost(self, cost: float) -> None:
-        """
-        Update the cost of the session.
-        """
-        if isinstance(cost, float) and isinstance(self._cost, float):
-            self._cost += cost
-        else:
-            self._cost = None
